+++ examples/relay/consumer.php
@@ -26,6 +26,18 @@ class CustomConsumer extends Consumer {
     print_r($message);
   }
 
+  public function onIncomingMessage($message): Coroutine {
+    yield;
+    echo "\n New inbound message \n";
+    print_r($message);
+  }
+
+  public function onMessageStateChange($message): Coroutine {
+    yield;
+    echo "\n Message state changed \n";
+    print_r($message);
+  }
+
   public function onIncomingCall($call): Coroutine {
     print "\n - onIncomingCall on context: {$call->context}, from: {$call->from} to: {$call->to} !\n";