+++ src/Relay/Tasking/Tasking.php
@@ -8,7 +8,7 @@ class Tasking extends \SignalWire\Relay\BaseRelay {
 
   public function notificationHandler($notification): void {
     Log::info("Receive task in context: {$notification->context}");
-    Handler::trigger($this->client->relayProtocol, $notification->message, $this->_prefixCtx($notification->context));
+    Handler::trigger($this->client->relayProtocol, $notification->message, $this->_ctxReceiveUniqueId($notification->context));
   }
 
 }